back/requirements.txt
alembic==1.8.1
autopep8==1.7.0
click==8.1.3
Faker==15.0.0
Flask==2.2.2
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
mysqlclient==2.1.1
packaging==21.3
pycodestyle==2.9.1
PyJWT==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
toml==0.10.2
Werkzeug==2.2.2